Extract modal/ sub-package from internal/tui
Move confirmdelete.go and confirmdeleteproject.go to
internal/tui/modal/. These are self-contained confirm dialogs
with no cross-package dependencies.

Phase 6 (partial) of the tui package decomposition.
