+++ internal/tui/modal/confirmdelete.go
@@ -1,4 +1,4 @@
-package tui
+package modal
 
 import (
 	"github.com/gdamore/tcell/v2"
