+++ internal/tui/modal/confirmdeleteproject.go
@@ -1,4 +1,4 @@
-package tui
+package modal
 
 import (
 	"fmt"
